requirements.txt
asgiref==3.8.1
Django==5.1.4
pip==23.2.1
sqlparse==0.5.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
gunicorn==20.1.0
